Move JSON editor into BaseDashModel
